Fixed issues with releasing the real primary surface
The palette converter surface was being released even when it was null,
leading to potential problems with releasing the primary surface.

The update thread is also no longer terminated on each release.
Termination may have been necessary more often than expected because
the thread may be waiting on the DirectDraw critical section on release.
There doesn't seem to be any issues with letting the OS clean up
the running thread on exit instead.