Fixed synchronization issue in beginGdiRendering
It's not sufficient to make g_renderingRefCount atomic, because after it is
found to be nonzero in beginGdiRendering, a call to endGdiRendering from
another thread may make it zero before beginGdiRendering has a chance of
entering the critical section. This would mean that the DirectDraw primary
surface is no longer locked, even though beginGdiRendering expects that it is.

To prevent potential issues, the g_renderingRefCount check is now also
guarded by the critical section.