Hook IDirect3DDevice7 only if needed
Creating an instance of the IDirect3DDevice7 interface appears to cause
problems in some games that use only older versions of Direct3D.
IDirect3DDevice7 is now only hooked when an application creates an instance.

Fixes video playback artifacts in Populous: The Beginning.
Fixes a crash in Carmageddon (Win95 version) reported in issue #3.