Restore primary surface globals if create fails
Fixes error creating "second" primary in Prince of Persia 3D when
starting a new game.  First primary is not completely released because
IDirectDrawGammaControl is not released. Native ddraw still allows
creation of a second fullscreen primary in this case.

This fix replaces the previous fix for Conquest: Frontier Wars also.